get_compass_image: Wrap in a while loop and retry if no screen data found
get_compass_image sometimes ends up with empty data in the compass_image and the screen array following the calculation trying to locate the navigation dot.  The causes an exception in filter_blue() and results in the autopilot failing. This seems to be a result of a bad screen capture providing garbage data, so if we just try again when we notice compass_image is empty, we always end up returning valid data.

ref: https://github.com/skai2/EDAutopilot/issues/36

Signed-off-by: Joe MacDonald <joe@deserted.net>